(Partitioner) Rename, refactor and add logging info
