+++ plugins/org.preesm.algorithm/src/org/preesm/algorithm/clustering/partitioner/ClusterPartitionerTask.java
@@ -1,4 +1,4 @@
-package org.preesm.algorithm.clustering.partionner;
+package org.preesm.algorithm.clustering.partitioner;
 
 import java.util.HashMap;
 import java.util.LinkedHashMap;
@@ -10,28 +10,26 @@ import org.preesm.commons.doc.annotations.PreesmTask;
 import org.preesm.commons.doc.annotations.Value;
 import org.preesm.model.pisdf.PiGraph;
 import org.preesm.model.scenario.Scenario;
-import org.preesm.model.slam.Design;
 import org.preesm.workflow.elements.Workflow;
 import org.preesm.workflow.implement.AbstractTaskImplementation;
 
 /**
- * Cluster Partionner Task
+ * Cluster Partitioner Task
  * 
  * @author dgageot
  *
  */
-@PreesmTask(id = "cluster-partionner", name = "Cluster Partionner",
+@PreesmTask(id = "cluster-partitioner", name = "Cluster Partitioner",
     inputs = { @Port(name = "PiMM", type = PiGraph.class, description = "Input PiSDF graph"),
-        @Port(name = "scenario", type = Scenario.class, description = "Scenario"),
-        @Port(name = "architecture", type = Design.class, description = "Architecture") },
+        @Port(name = "scenario", type = Scenario.class, description = "Scenario") },
     outputs = { @Port(name = "PiMM", type = PiGraph.class, description = "Output PiSDF graph") },
     parameters = { @Parameter(name = "Number of PEs in clusters",
-        description = "The number of PEs in clusters which is used to balance firings upon coarse-grained "
-            + "and fine-grained levels.",
+        description = "The number of PEs in compute clusters. This information is used to balance actor firings"
+            + " between coarse and fine-grained levels.",
         values = { @Value(name = "Fixed:=n", effect = "Where $$n\\in \\mathbb{N}^*$$.") }) })
-public class ClusterPartionnerTask extends AbstractTaskImplementation {
+public class ClusterPartitionerTask extends AbstractTaskImplementation {
 
-  public static final String NB_PE         = "Number of PEs in clusters";
+  public static final String NB_PE         = "Number of PEs in compute clusters";
   public static final String DEFAULT_NB_PE = "1";
 
   @Override
@@ -40,12 +38,12 @@ public class ClusterPartionnerTask extends AbstractTaskImplementation {
     // Task inputs
     PiGraph inputGraph = (PiGraph) inputs.get("PiMM");
     Scenario scenario = (Scenario) inputs.get("scenario");
-    Design architecture = (Design) inputs.get("architecture");
 
     // Parameters
     String nbPE = parameters.get(NB_PE);
 
-    PiGraph outputGraph = new ClusterPartionner(inputGraph, scenario, architecture, Integer.parseInt(nbPE)).cluster();
+    // Cluster input graph
+    PiGraph outputGraph = new ClusterPartitioner(inputGraph, scenario, Integer.parseInt(nbPE)).cluster();
 
     // Build output map
     Map<String, Object> output = new HashMap<>();
@@ -63,7 +61,7 @@ public class ClusterPartionnerTask extends AbstractTaskImplementation {
 
   @Override
   public String monitorMessage() {
-    return "Starting Execution of Cluster Partionner Task";
+    return "Starting Execution of Cluster Partitioner Task";
   }
 
 }
